fix!: Display correct chart data for logged in users
Issue was that the /feature-flags data was not available during while
components were being constructed, when user was logged in. Not sure
why, but removing the flag check in the constructor solved the issue.
There was also a workaround that a logged in user could select a
different chart and come back to the Developers chart, and they'd see
the right data

While fixing this issue, I also removed the ERD+3 months flag from
the charts ticket, as required for OCD-3532

[#OCD-3539]