test[aqa]: Add c3 CQM error that is now given appropriately by the API
The listing related to the added error message attests to both
c3-original and c3-Cures. It has a single CQM that references "c3". "c3"
is interpreted by the API to mean c3-Cures, so it is a correct error
that the listing attests to c3-original but does not have any related
CQMs.

[#OCD-4071]