fix: Display job edit screen when there are no subscribers
I removed Ashwini's hard-coded email address as a Trigger Developer Ban
subscriber with OCD-4060. When the QA team had to do a reset of the
quartz jobs to regression test functionality related to Direct Reviews,
the removal of Ashwini's subscription finally took effect and the
Trigger Developer Ban job could not be edited when the email list was
'undefined' (cannot 'split' 'undefined'). I added some simple logic
to use an empty array in that case and to use the provided emails
otherwise.

[#OCD-4155]