fix!: Check dev id instead of dev name in product owner history
There are a few examples in the data where a product has an owner
history and the previous owner has the same name as the current owner
but upon inspection of the data they are actually different developers.
See developer ID 369. This seems due to joins and renames of developers
over time.