test[aqa]: CSS Selector specifies 'svg' element one level deep
A new 'svg' type of element was added in the HTML and was being
selected. The element being selected was no longer the "x" button to
remove a filter but instead was the new icon indicating certification
status. The more specific 'svg element one level deep' once again select
the correct element with the "x".